Add files via upload
This project first calculated the redshift of a number of spectra which in turn was used to correct for time dilation when analysing light curves in the B-band. Peak magnitude in the B-band was found, followed by the light curve width $\Delta m_{15}(B)$. A Hubble diagram was created using the light curve widths and peak absolute magnitudes which found a luminosity-decline rate relation. Finally, three different models are explored in order to deduce the most accurate in calculating distance modulus.